wip: add vcpu, vm files
