+++ vcpu.h
@@ -1,5 +1,10 @@
 #pragma once
+
 #include <linux/kvm.h>
+#include <sys/mman.h>
+#include <sys/ioctl.h>
+#include "vm.h"
+#include "util.h"
 
 struct vcpu {
     int fd;
@@ -7,3 +12,7 @@ struct vcpu {
     struct kvm_sregs sregs;
     struct kvm_regs regs;
 };
+
+void init_vcpu(struct vm *vm, struct vcpu *vcpu);
+void set_sregs(struct kvm_sregs *sregs);
+void set_regs(struct vcpu *vcpu);
