+++ vm.c
@@ -1,0 +1,29 @@
+#include "vm.h"
+
+void create_vm(struct vm *vm) {
+    vm->fd = ioctl(vm->vm_fd, KVM_CREATE_VM, 0);
+    if (vm->fd < 0) {
+        error("KVM_CREATE_VM");
+    }
+}
+
+void set_tss(int fd) {
+    if (ioctl(fd, KVM_SET_TSS_ADDR, 0xfffbd000) < 0) {
+        error("KVM_SET_TSS_ADDR");
+    }
+}
+
+void load_guest_binary(void *dst) {
+    int fd = open("../xv6/bootblock", O_RDONLY);
+    if (fd < 0) 
+        error("open fail");
+    
+    void *tmp = dst;
+    for(;;) {
+        int size = read(fd, tmp + START_ADDRESS, 
+                        GUEST_BINARY_SIZE);
+        if (size <= 0) 
+            break;
+        tmp += size;
+    }
+}
